+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixed a bug with loading 2.X on a save that played with 1.0
 * All Lua-based strings including console commands are now localized
 * More safeguards around code that checks if the player is in combat or not
+* Travel points will now only be generated if the player is on solid ground
 
 <!-- [Download Link](https://gitlab.com/modding-openmw/signpost-fast-travel/-/packages/#TODO) -->